Split out paragraph marker logic
